+++ src/inspect_ai/_view/www/src/samples/scores/SampleScores.tsx
@@ -1,8 +1,7 @@
-import { FC, Fragment } from "react";
+import { FC } from "react";
 import { SampleSummary } from "../../api/types";
 
-import { useSampleDescriptor } from "../../state/hooks";
-import styles from "./SampleScores.module.css";
+import { getScoreDescriptorForValues } from "../descriptor/score/ScoreDescriptor";
 
 interface SampleScoresProps {
   sample: SampleSummary;
@@ -10,24 +9,14 @@ interface SampleScoresProps {
 }
 
 export const SampleScores: FC<SampleScoresProps> = ({ sample, scorer }) => {
-  const samplesDescriptor = useSampleDescriptor();
-  const scores = scorer
-    ? samplesDescriptor?.evalDescriptor
-        .scorerDescriptor(sample, { scorer, name: scorer })
-        .scores()
-    : samplesDescriptor?.selectedScorerDescriptor(sample)?.scores();
-
-  if (scores?.length === 1) {
-    return scores[0].rendered();
-  } else {
-    const rows = scores?.map((score) => {
-      return (
-        <Fragment>
-          <div style={{ opacity: "0.7" }}>{score.name}</div>
-          <div>{score.rendered()}</div>
-        </Fragment>
-      );
-    });
-    return <div className={styles.grid}>{rows}</div>;
+  const scoreData = sample.scores?.[scorer];
+  if (!scoreData) {
+    return undefined;
   }
+
+  const scorerDescriptor = getScoreDescriptorForValues(
+    [scoreData.value],
+    [typeof scoreData.value],
+  );
+  return scorerDescriptor?.render(scoreData.value);
 };