+++ CHANGELOG.md
@@ -1,7 +1,36 @@
+## Unreleased
+
+- [Agent](https://inspect.aisi.org.uk/agent-protocol.html) protocol and [inspect_ai.agent](https://inspect.aisi.org.uk/reference/inspect_ai.agent.html) module with new system for creating, composing, and executing agents.
+- Model API: New `execute_tools()` function (replaces deprecated `call_tools()` function) which handles agent handoffs that occur during tool calling.
+- Model API: `generate_loop()` method for calling generate with a tool use loop.
+- Scoring: New `grouped` metric, which computes a metric for a list of scores grouped by a value in sample metadata.
+- Tools: `bash_session()` and `web_browser()` now create a distinct sandbox process each time they are instantiated.
+- Task API: `task_with()` and `tool_with()` no longer copy the input task or tool (rather, they modify it in place and return it).
+- Eval Set: Resolve tasks before each pass (ensure that each pass runs against an entirely new task instance).
+- Eval Retry: Ability to retry any task in the registry, even if it has a custom `name` (save `registry_name` separately).
+- Human Agent: Start task with clock paused and then automatically start it on container logins.
+- Typed Store: `instance` option for `store_as()` for using multiple instances of a `StoreModel` within a sample.
+- Typed Store: Raise error if attempting to embed a `StoreModel` within another `StoreModel`.
+- Docker: `write_file()` function now gracefully handles larger input file sizes (was failing on files > 2MB).
+- Docker: Prevent low timeout values (e.g. 1 second) from disabling timeout entirely when they are retried.
+- Inspect View: Live updates to running evaluation logs.
+- Inspect View: Fallback to content range request if inital HEAD request fails.
+- Inspect View: Improve error message when view bundles are server from incompatible servers.
+- Inspect View: Render messages in `user` and `assistant` solver events.
+- Inspect View: Improved support for display of nested arrays.
+- Inspect View: Improved rendering of complex scores and metrics.
+- Inspect View: Properly handle filtering of dictionary scores.
+- Inspect View: Render math in model input and output using katex.
+- Inspect View: Improve sample score rendering (single scoring tab with scores rendered in a table).
+- Bugfix: Support for calling the `score()` function within Jupyter notebooks.
+- Bugfix: Handle process lookup errors that can occur during timeout race conditions.
+- Bugfix: Correctly capture and return logs from `eval()` when a cancellation occurs.
+- Bugfix: Correctly handle custom `api_version` model argument for OpenAI on Azure.
+
 ## v0.3.82 (02 April 2025)
 
 - Bugfix: Correct handling of backward compatiblity for inspect-web-browser-tool image.
-- Bugfix: Eval now properly exits when `max_tasks` is greater than total tasks.
+- Bugfix: Eval now properly exits when `max_tasks` is greater than total tasks
 
 ## v0.3.81 (30 March 2025)
 
@@ -12,14 +41,6 @@
 - Google: Compatibility with httpx client in `google-genai` >= 1.8.0 (which is now required).
 - Mistral: Compatibility with tool call schema for `mistralai` >= v1.6.0 (which is now required).
 - Inspect View: Correctly parse NaN values (use JSON5 for all JSON parsing)
-
-## Unreleased
-
-- Inspect View: Live updates to running evaluation logs.
-- Inspect View: Fallback to content range request if inital HEAD request fails.
-- Inspect View: Improve error message when view bundles are server from incompatible servers.
-- Inspect View: Render messages in `user` and `assistant` solver events.
-- Inspect View: Improved support for display of nested arrays.
 
 ## v0.3.79 (26 March 2025)